# 임포트

+++ src/main/java/egovframework/srv/response/service/ResponseService.java
@@ -7,6 +7,8 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import egovframework.rte.psl.dataaccess.util.EgovMap;
+import egovframework.srv.management.dao.IManagementDAO;
 import egovframework.srv.response.dao.IResponseDAO;
 
 @Service
@@ -14,6 +16,9 @@ public class ResponseService{
 
 	@Autowired
 	private IResponseDAO dao;
+	
+	@Autowired
+	private IManagementDAO managementDao;
 	
 	@SuppressWarnings("unchecked")
 	public void insert(Map<String,Object> map) throws Exception {
@@ -23,10 +28,39 @@ public class ResponseService{
 			dao.insert(inputMap);
 		}
 	}
-	public void srvCount(Map<String,Object> map) throws Exception {
+	public String srvCount(Map<String,Object> map) throws Exception {
 		Integer srvMstrId = (Integer) map.get("srvMstrId");
-		dao.deleteRespnstt(srvMstrId);
-		dao.insertRespnstt(srvMstrId);
+		
+		Integer countSrvRespns = dao.countSrvRespns(srvMstrId);
+		if( countSrvRespns >  0) {
+			dao.deleteRespnstt(srvMstrId);
+			dao.insertRespnstt(srvMstrId);
+			return "0";
+		}else {
+			return "1";
+		}
+	}
+	
+	public EgovMap showStt(Map<String,Object> map) throws Exception {
+		Integer srvMstrId = Integer.parseInt( map.get("srvMstrId").toString() );
+		
+		HashMap<String, Object> inputMap = new HashMap<String,Object>();
+		inputMap.put("value", srvMstrId);
+		EgovMap srvMstr = managementDao.selectSrvMstr(inputMap);
+		List<EgovMap> srvMtrItemsMap = dao.selectSttObjLabel(srvMstrId);
+		
+		for (EgovMap egovMap : srvMtrItemsMap) {
+			inputMap = new HashMap<String,Object>();
+			inputMap.put("srvMstrId", srvMstrId);
+			inputMap.put("srvMtritemsId", Integer.parseInt(egovMap.get("srvMtritemsId").toString()));
+			List<EgovMap> selectSttObjData = dao.selectSttObjData(inputMap);
+			egovMap.put("sttObjData", selectSttObjData);
+		}
+		
+		EgovMap resultMap = new EgovMap(); 
+		resultMap.put("srvMstr",srvMstr);
+		resultMap.put("srvMtrItems", srvMtrItemsMap);
+		return resultMap;
 	}
 
 }
